DI – remove gif repo, refactoring
