+++ app/src/main/java/com/alex/eyk/gifsearch/data/net/exception/EmptyResponseBodyException.kt
@@ -1,0 +1,3 @@
+package com.alex.eyk.gifsearch.data.net.exception
+
+class EmptyResponseBodyException : Exception()
