+++ app/src/main/java/com/alex/eyk/gifsearch/data/net/exception/ServerException.kt
@@ -1,0 +1,3 @@
+package com.alex.eyk.gifsearch.data.net.exception
+
+class ServerException : Exception()
