--- app/src/main/java/com/alex/eyk/gifsearch/data/repo/GifRepository.kt
@@ -1,13 +1,0 @@
-package com.alex.eyk.gifsearch.data.repo
-
-import com.alex.eyk.gifsearch.Either
-import com.alex.eyk.gifsearch.data.entity.Gif
-
-interface GifRepository {
-
-    suspend fun findAll(
-        query: String,
-        limit: Int,
-        offset: Int
-    ): Either<List<Gif>>
-}
