--- app/src/main/java/com/alex/eyk/gifsearch/data/repo/exception/ServerException.kt
@@ -1,3 +1,0 @@
-package com.alex.eyk.gifsearch.data.repo.exception
-
-class ServerException : Exception()
