+++ app/src/main/java/com/alex/eyk/gifsearch/data/repo/impl/SuggestionsRemoteRepository.kt
@@ -2,6 +2,7 @@ package com.alex.eyk.gifsearch.data.repo.impl
 
 import com.alex.eyk.gifsearch.Either
 import com.alex.eyk.gifsearch.data.entity.Suggestion
+import com.alex.eyk.gifsearch.data.net.handleResponse
 import com.alex.eyk.gifsearch.data.net.service.GifService
 import com.alex.eyk.gifsearch.data.repo.SuggestionsRepository
